refactor: run dufs in blocking process mode
Most folks will use the http server for a short time, and then switch back to their apps. Rather than run in the background and consume resources, run it in process mode by default. Daemon-mode can be set by creating a file named "daemon-mode" in the pak folder.

This works by using evtest - an input device tester - to check if the B button is pressed and then exiting appropriately. Logs are now written to a single log directory with some rudimentary log rotation (keeps the runs of the last execution) to reduce disk utilization.